Move navBar2 functionality to navBar, delete navBar2
